Add links to the referenced protocols
In order to reading up the spec of a given protocol easier, the list of
available protocols is converted from an ordinary protocol-to-overhead
map into a protocol-to-protocol-object map. The protocol-object consists
of the name, the overhead and an optional link to the protocol. That
link is rendered in the view of available protocols.

Also part of this commit:

* The previous rendering of the available protocols used `<input>`
  buttons. They can not carry hyperlinks. That is why they are now
  ordinary `<div>` elements.

* Getting rid of the "serial-number-hack": Using a DOM element to carry
  a monotonic counter is a bit too heavy when a local counter will do.
  The `generate_id()` function will create the equivalent to how the
  DOM based serial generation worked before.

* Several smaller code reorderings / renamings to prepare code
  refactoring to introduce a proper "model" which holds the state and
  then using function to render that state. Currently, the state of
  the selected protocol-stack is encoded in the DOM.